added settings button for new items
