+++ frontend/src/styles/card-style.css
